Move `pages/mocks` to `mocks/`
