+++ x-pack/plugins/security_solution/public/management/components/artifact_list_page/components/artifact_delete_modal.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { AppContextTestRender } from '../../../../common/mock/endpoint';
-import { trustedAppsAllHttpMocks } from '../../../pages/mocks';
+import { trustedAppsAllHttpMocks } from '../../../mocks';
 import {
   ArtifactListPageRenderingSetup,
   getArtifactListPageRenderingSetup,
