+++ x-pack/plugins/security_solution/public/management/components/artifact_list_page/mocks.tsx
@@ -13,7 +13,7 @@ import userEvent from '@testing-library/user-event';
 import { ArtifactFormComponentProps } from './types';
 import { ArtifactListPage, ArtifactListPageProps } from './artifact_list_page';
 import { AppContextTestRender, createAppRootMockRenderer } from '../../../common/mock/endpoint';
-import { trustedAppsAllHttpMocks } from '../../pages/mocks';
+import { trustedAppsAllHttpMocks } from '../../mocks';
 import { TrustedAppsApiClient } from '../../pages/trusted_apps/service/api_client';
 import { artifactListPageLabels } from './translations';
