+++ x-pack/plugins/security_solution/public/management/components/endpoint_console/isolate_action.test.tsx
@@ -12,7 +12,7 @@ import {
 } from '../console/components/console_manager/mocks';
 import React from 'react';
 import { getEndpointResponseActionsConsoleCommands } from './endpoint_response_actions_console_commands';
-import { responseActionsHttpMocks } from '../../pages/mocks/response_actions_http_mocks';
+import { responseActionsHttpMocks } from '../../mocks/response_actions_http_mocks';
 import { enterConsoleCommand } from '../console/mocks';
 import { waitFor } from '@testing-library/react';
