+++ x-pack/plugins/security_solution/public/management/hooks/endpoint/use_get_action_details.test.ts
@@ -11,7 +11,7 @@ import {
   ReactQueryHookRenderer,
 } from '../../../common/mock/endpoint';
 import { useGetActionDetails } from './use_get_action_details';
-import { responseActionsHttpMocks } from '../../pages/mocks/response_actions_http_mocks';
+import { responseActionsHttpMocks } from '../../mocks/response_actions_http_mocks';
 import { resolvePathVariables } from '../../../common/utils/resolve_path_variables';
 import { ACTION_DETAILS_ROUTE } from '../../../../common/endpoint/constants';
 import { useQuery as _useQuery } from 'react-query';
