+++ x-pack/plugins/security_solution/public/management/hooks/endpoint/use_get_endpoint_pending_actions_summary.test.ts
@@ -13,7 +13,7 @@ import {
 import { useGetEndpointPendingActionsSummary } from './use_get_endpoint_pending_actions_summary';
 import { ACTION_STATUS_ROUTE } from '../../../../common/endpoint/constants';
 import { useQuery as _useQuery } from 'react-query';
-import { responseActionsHttpMocks } from '../../pages/mocks/response_actions_http_mocks';
+import { responseActionsHttpMocks } from '../../mocks/response_actions_http_mocks';
 
 const useQueryMock = _useQuery as jest.Mock;
