+++ x-pack/plugins/security_solution/public/management/mocks/index.ts
