+++ x-pack/plugins/security_solution/public/management/mocks/trusted_apps_http_mocks.ts
@@ -26,8 +26,8 @@ import {
   composeHttpHandlerMocks,
   httpHandlerMockFactory,
   ResponseProvidersInterface,
-} from '../../../common/mock/endpoint/http_handler_mock_factory';
-import { ExceptionsListItemGenerator } from '../../../../common/endpoint/data_generators/exceptions_list_item_generator';
+} from '../../common/mock/endpoint/http_handler_mock_factory';
+import { ExceptionsListItemGenerator } from '../../../common/endpoint/data_generators/exceptions_list_item_generator';
 import {
   fleetGetAgentPolicyListHttpMock,
   FleetGetAgentPolicyListHttpMockInterface,
@@ -37,7 +37,7 @@ import {
 import {
   BY_POLICY_ARTIFACT_TAG_PREFIX,
   GLOBAL_ARTIFACT_TAG,
-} from '../../../../common/endpoint/service/artifacts/constants';
+} from '../../../common/endpoint/service/artifacts/constants';
 
 interface FindExceptionListItemSchemaQueryParams
   extends Omit<FindExceptionListItemSchema, 'page' | 'per_page'> {
