+++ x-pack/plugins/security_solution/public/management/pages/event_filters/view/event_filters_list.test.tsx
@@ -11,7 +11,7 @@ import React from 'react';
 import { EVENT_FILTERS_PATH } from '../../../../../common/constants';
 import { AppContextTestRender, createAppRootMockRenderer } from '../../../../common/mock/endpoint';
 import { EventFiltersList } from './event_filters_list';
-import { exceptionsListAllHttpMocks } from '../../mocks/exceptions_list_http_mocks';
+import { exceptionsListAllHttpMocks } from '../../../mocks/exceptions_list_http_mocks';
 import { SEARCHABLE_FIELDS } from '../constants';
 import { parseQueryFilterToKQL } from '../../../common/utils';
