+++ x-pack/plugins/security_solution/public/management/pages/host_isolation_exceptions/view/components/form.test.tsx
@@ -21,7 +21,7 @@ import { HOST_ISOLATION_EXCEPTIONS_PATH } from '../../../../../../common/constan
 import {
   exceptionsListAllHttpMocks,
   fleetGetEndpointPackagePolicyListHttpMock,
-} from '../../../mocks';
+} from '../../../../mocks';
 import {
   clickOnEffectedPolicy,
   isEffectedPolicySelected,
