+++ x-pack/plugins/security_solution/public/management/pages/policy/view/ingest_manager_integration/endpoint_package_custom_extension/endpoint_package_custom_extension.test.tsx
@@ -11,7 +11,7 @@ import { createFleetContextRendererMock, generateFleetPackageInfo } from '../moc
 import { EndpointPackageCustomExtension } from './endpoint_package_custom_extension';
 import { useEndpointPrivileges as _useEndpointPrivileges } from '../../../../../../common/components/user_privileges/endpoint/use_endpoint_privileges';
 import { getEndpointPrivilegesInitialStateMock } from '../../../../../../common/components/user_privileges/endpoint/mocks';
-import { exceptionsListAllHttpMocks } from '../../../../mocks/exceptions_list_http_mocks';
+import { exceptionsListAllHttpMocks } from '../../../../../mocks/exceptions_list_http_mocks';
 import { waitFor } from '@testing-library/react';
 
 jest.mock('../../../../../../common/components/user_privileges/endpoint/use_endpoint_privileges');
