+++ x-pack/plugins/security_solution/public/management/pages/policy/view/policy_details.test.tsx
@@ -17,7 +17,7 @@ import { policyListApiPathHandlers } from '../store/test_mock_utils';
 import { PolicyDetails } from './policy_details';
 import { APP_UI_ID } from '../../../../../common/constants';
 import { useIsExperimentalFeatureEnabled } from '../../../../common/hooks/use_experimental_features';
-import { exceptionsFindHttpMocks } from '../../mocks/exceptions_list_http_mocks';
+import { exceptionsFindHttpMocks } from '../../../mocks/exceptions_list_http_mocks';
 
 jest.mock('./policy_forms/components/policy_form_layout');
 jest.mock('../../../../common/components/user_privileges');
