+++ x-pack/plugins/security_solution/public/management/pages/trusted_apps/view/trusted_apps_list.test.tsx
@@ -11,7 +11,7 @@ import React from 'react';
 import { TRUSTED_APPS_PATH } from '../../../../../common/constants';
 import { AppContextTestRender, createAppRootMockRenderer } from '../../../../common/mock/endpoint';
 import { TrustedAppsList } from './trusted_apps_list';
-import { exceptionsListAllHttpMocks } from '../../mocks/exceptions_list_http_mocks';
+import { exceptionsListAllHttpMocks } from '../../../mocks/exceptions_list_http_mocks';
 import { SEARCHABLE_FIELDS } from '../constants';
 import { parseQueryFilterToKQL } from '../../../common/utils';
